# api/swagger.yaml
swagger: '2.0'
info:
  version: "1.0.1"
  title: teacher-admin-apis
  description: This is a set of endpoints for teachers to perform administrative functions for their classes.
host: localhost:3000
schemes:
  - http
  - https
consumes:
  - application/json
produces:
  - application/json
basePath: /
paths:
...
  /api/register:
    post:
      summary: Register Students
      description: |
         Register one or more students to a specified teacher.
      operationId: "registerStud"   
      parameters:
        - name: teacher 
          in: body
          description: register students
          required: true
          schema:
            $ref: '#/definitions/Teacher'
          
      responses:
        200:
          description: Successful response
          schema:
              title: Return String
              type: string
              example: "Succesfully Registered"
        500:
          description: Error
          schema: 
            type: string
            example: "Could not register Students"
      x-swagger-router-controller: "Default"      
...
definitions:
  Teacher:
    description: Teacher object
    properties:
      teacher:
        type: string
        description: teacher email
      students:
        type: array
        description: add students
        items:
          $ref: '#/definitions/Teacher'
    required:
      - teacher
      - students
...